Polish
added rotation, answering and parameters with desciptions.
