+++ Assets/Core (nix pille!)/Scripts/SwipeInput.cs.meta
